requirements.txt
discord.py
google-genai
flask
gunicorn